correcting price to be whole numbers only

+++ db/migrations/20180810152501_snacks.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
       table.increments()
       table.string('name').notNullable().defaultTo('')
       table.text('description').notNullable().defaultTo('')
-      table.float('price')
+      table.integer('price')
       table.text('img').notNullable().defaultTo('')
       table.boolean('is_perishable').notNullable().defaultTo(false)
     })
